refactor: typescript alias
