+++ backend/package.json
@@ -2,7 +2,7 @@
   "name": "repo-be",
   "version": "0.0.1",
   "scripts": {
-    "start": "nodemon --watch 'src/**/*.ts' --exec 'ts-node' src/main.ts",
+    "start": "nodemon --watch 'src/**/*.ts' --exec 'ts-node' -r tsconfig-paths/register src/main.ts",
     "serve": "rm -rf dist && yarn tsc && node dist/main"
   },
   "dependencies": {
@@ -22,6 +22,7 @@
     "eslint-plugin-import": "^2.26.0",
     "nodemon": "^2.0.16",
     "ts-node": "^10.8.0",
+    "tsconfig-paths": "^4.0.0",
     "typescript": "^4.7.2"
   }
 }
